Moved the realized element list outside.
Class was nearly as big as `VirtualizingStackPanel` itself and getting a bit unwieldy.
